Move all new effects into a separate directory
