+++ packages/flame/lib/src/effects_new/effect_component.dart
