+++ packages/flame/lib/src/effects_new/flame_animation_controller.dart
