+++ packages/flame/lib/src/effects_new/standard_animation_controller.dart
